requirements.txt
fastapi>=0.109.0
python-dotenv>=1.0.0
openai>=1.10.0
httpx>=0.26.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
uvicorn[standard]>=0.27.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
duckduckgo-search>=4.0.0
newspaper3k>=0.2.8
readability-lxml>=0.8.1
selenium>=4.15.0
requests>=2.31.0
aiofiles>=23.0.0

